backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
groq==0.4.1
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
